--- generic-exporter-example/src/main/resources/pre_transformer.json
@@ -1,12 +1,0 @@
-{
-    "transformations": [
-        {},
-        {
-            "sourcePointer": "/datasetVersion/metadataBlocks/citation/fields",
-            "resultPointer": "/datasetVersion/metadataBlocks/citation",
-            "expressions": [
-                "script(importJS /js/map_metadata_fields.js endImport)"
-            ]
-        }
-    ]
-}
